% Generated by roxygen2 (4.0.0): do not edit by hand
\name{calcW}
\alias{calcW}
\title{Calculate David's score 'w' paramter}
\usage{
calcW(player, intData)
}
\arguments{
\item{player}{in the "interData" object of which to find the 'w' parameter}

\item{intData}{an object of class "interData" which contains the players
interactions.}
}
\value{
w a sum of individuals Pij percentages.
}
\description{
Calculates the w paramter of the Davids score algorithim. W is
calculated as the sum of all individuals Pij percentages where i is the
individual for which w is associated and and j is all other players in the
interData object. Ties are disregarded in this
calculation as per the methods described in Gammel et al. 2003
}
\details{
Find the 'w' paramter of the David's score algorithim for the associated
player defined in the player argument
}
\references{
Gammel et al. (2003) David's Score. Animal Behaviour
}

