% Generated by roxygen2 (4.0.0): do not edit by hand
\name{calcWL2}
\alias{calcWL2}
\title{Calculate David's score 'w2' or 'l2' score}
\usage{
calcWL2(player, intData, param = "w")
}
\arguments{
\item{player}{in the "interData" object of which to find the 'w' parameter}

\item{intData}{an object of class "interData" which contains the players
interactions.}

\item{param}{either the single character w or l to calculate the
w2 or l2 parmeter respectively.}
}
\value{
Either w2 or l2 as described above.
}
\description{
Calculates the w2 or l2 score of the Davids score
algorithim. w2 is calculated as the sum of all individuals Pij percentages
where i is the individual for which w is associated and and j is all other
players in the interData object weighted by player j's w score. l2 is the
sum of all (1 - Pij) weighted by player j's l score. Ties are disregarded
in this calculation as per the methods described in Gammel et al. 2003
}
\details{
Find the 'w2' or 'l2' score of the David's score algorithim for the
associated player defined in the player argument.
}
\examples{
# generate generic data
interactions <- data.frame (a = sample (letters [1:10], 100, T),
                            b = sample (letters [1:10], 100, T),
                            o = sample (c(-1,-1,0,1,1), 100, T),
                            d = Sys.time () + runif (100, 40, 160))
# convert to interData object
id1 <- intTableConv (interactions)
# find w2 of player in group
calcWL2 (id1$players [1], id1)
# find l2 of player in group
calcWL2 (id1$players [1], id1, 'l')
}
\references{
Gammel et al. (2003) David's Score. Animal Behaviour
}

