% Generated by roxygen2 (4.0.0): do not edit by hand
\name{eloTable}
\alias{eloTable}
\title{Generate a series of elo ratings from interaction data}
\usage{
eloTable(interData, initial = 1000, ...)
}
\arguments{
\item{interData}{an onject of class "interData".}

\item{initial}{the starting score for all participants in the group.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
An object of the class eloTable which contains a list of players
in the graoup, the date range of interactions and a data frame sorted by
date containing any changes in an individuals elo rating.
}
\description{
Given an object of the class "interData" will produce a series of elo
ratings up until the latest interaction in the interData object.
}
\details{
Using the methods described in Neumann et al 2011. a series of Elo
ratings are generated from interaction data. Initial scores may either be a
single numeric value or a vector of integers correspong to the players
element of interData. Further arguments are passed to the findK and
singleOutcome functions and include the paramters; kScaleBounds, kFactors,
and minTresh.
}
\references{
Neumann et al (2011) Assessing Dominance Hierarchies.
Animal Behaviour
}

