% Generated by roxygen2 (4.0.0): do not edit by hand
\name{extractScores}
\alias{extractScores}
\title{Get the latest scores from an eloTable object until specified time}
\usage{
extractScores(eloTab, tObj = Sys.time(), players = eloTab[["players"]])
}
\arguments{
\item{eloTab}{an object of the class "eloTable" from which ratings are to be
extracted.}

\item{tObj}{an element of the class "POSIXct" for which
only scores occuring prior to within the eloTable will be used.}

\item{players}{A character list of players for which scores are obtained}
}
\value{
A data frame containing the latest rating for each player up to the
datetime parameter as well as the time which that rating was last updated.
}
\description{
Pulls players latest scores from an eloTable object until specified time. If
no scores exist for an individual prior to that time in the table then the
initial value is returned.
}
\details{
Using the datetime parameter as a cieling for elo ratings this
function will return the latest rating for each player in the eloTable
object which has occured on or before that time. If no score has occured
prior to the specified time than the initial score is returned for that
player with a value of NA for the time.
}
\examples{
# generate generic data
interactions <- data.frame (p1 = c('i', 'j', 'i'), p2 = c('j', 'h', 'h'),
                            o = 1, d = c ('1/1/89', '1/3/89', '1/2/89'))
# convert to interData object
id1 <- intTableConv (interactions, format = '\%m/\%d/\%y')
# produce eloTable object
et1 <- eloTable (id1)
# get latest scores
extractScores (et1)
# only scores before '1/2/89'
extractScores (et1, tObj = as.POSIXct ('1/2/89', format = '\%m/\%d/\%y'))
}

