% Generated by roxygen2 (4.0.0): do not edit by hand
\name{findK}
\alias{findK}
\title{Generate the appropriate K factor given a players score}
\usage{
findK(playerScore, kScaleBounds = c(-Inf, Inf), kFactors = 100)
}
\arguments{
\item{playerScore}{the score of the player in question}

\item{kScaleBounds}{the player score bins to associate with K factors}

\item{kFactors}{the possible K factors to be used}
}
\value{
the players associated K factor
}
\description{
Given a players score, ranges of scores to associate with a K factor and the
K values to use for the given ranges the function will find the appropriate
K score to use for an elo rtaing update based on Neumann et al 2011.
}
\details{
Given a parameterization of kScaleBounds = c(-Inf, 2100, 2400, Inf)
and kFactors = c(32,24,16):

- If an individuals score is less than 2100 the kFactor assigned is 32

- Between 2100 and 2400 the kFactor assigned is 24

- A score above 2400 the the kFactor assigned is 16
}
\note{
In order to ensure a proper function call the kScaleBounds
parameter must be sorted from least to greatest and must be of a length
1 greater than the kFactors parameter. In addition assigning the first
element of kScaleBounds to -Inf and the last to Inf will allow the
players score to always remain within an acceptable range.
}
\references{
Neumann et al (2011) Assessing Dominance Hierarchies.
Animal Behaviour
}

