% Generated by roxygen2 (4.0.0): do not edit by hand
\name{idError}
\alias{idError}
\title{Calls error for class non-conformity}
\usage{
idError(obj, class = "interData")
}
\arguments{
\item{obj}{object to test class of.}

\item{class}{the class for which the object is tested}
}
\description{
Error if object not of specified class
}

