% Generated by roxygen2 (4.0.1): do not edit by hand
\name{numInt}
\alias{numInt}
\title{Count the number of interactions}
\usage{
numInt(p1, p2, intData, includeDraws = F, showInteractions = F)
}
\arguments{
\item{p1}{a player in the interData object}

\item{p2}{a player in the interData object}

\item{intData}{an object of the class "interData" with players p1 and p2}

\item{includeDraws}{logical indicating wether to include draws or not}

\item{showInteractions}{logical indicating wether to return a data frame
showing the interactions. Defaults to showing the number of interactions.}
}
\value{
Either the number of interactions between p1 and p2 in intData or a
dataframe displaying those interactions
}
\description{
Find the number of interactions occuring between two players in an interData
object
}
\examples{
# generate generic data
interactions <- data.frame (a = sample (letters [1:10], 100, T),
                            b = sample (letters [1:10], 100, T),
                            o = sample (c(-1,-1,0,1,1), 100, T),
                            d = Sys.time () + runif (100, 40, 160))
# convert to interData object
id1 <- intTableConv (interactions)
# calculate using all players
numInt (id1$players [1], id1$players [2], id1)
}

