% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.eloTable}
\alias{plot.eloTable}
\title{Plot and eloTable object}
\usage{
\method{plot}{eloTable}(eloTab, start.time = eloTab[[2]][1],
  end.time = eloTab[[2]][2], players = eloTab[[1]],
  pcol = 1:length(eloTab[[1]]), ...)
}
\arguments{
\item{eloTab}{object of the class "eloTable"}
}
\description{
Using the player elo ratings of an "eloTable" object, plots a time series
graph showing changes in individuals scores. Specified players must have
at least one score in the specified time range.
}

