% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plyrError}
\alias{plyrError}
\title{Calls error if player not in interData object}
\usage{
plyrError(players, intData)
}
\arguments{
\item{players}{proposed players to use for function}

\item{intData}{an object of class "interData"}
}
\description{
Error if any player in players not in "interData" object
}

