% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.interData}
\alias{print.interData}
\title{Print call for interData object}
\usage{
\method{print}{interData}(interData)
}
\arguments{
\item{interData}{an object of the class "interData"}
}
\description{
Prints each item in interData list and head of the interData data frame.
}
\details{
Prints only the head of the interData data.frame
}
\examples{
# generate generic data
interactions <- data.frame (a = sample (letters [1:10], 100, T),
                            b = sample (letters [1:10], 100, T),
                            o = sample (c(-1,-1,0,1,1), 100, T),
                            d = Sys.time () + runif (100, 40, 160))
# convert to interData object
id1 <- intTableConv (interactions)
# print
print (id1)
}

