% Generated by roxygen2 (4.0.0): do not edit by hand
\name{singleOutcome}
\alias{singleOutcome}
\title{Generate New Elo Scores from a Single Dyadic Interaction}
\usage{
singleOutcome(R1, R2, out, k1, k2, minThresh = 100, ...)
}
\arguments{
\item{R1}{Current score of player.1.}

\item{R2}{Current score of player.2.}

\item{out}{either 1 player 1 wins, -1 player 2 wins, or 0 tie}

\item{kA}{the k factor for player 1}

\item{kB}{the k factor for player 2}

\item{minThresh}{the minimum value of any players score}
}
\value{
a vector with player 1's and 2's new scores
}
\description{
Takes the elo score of two individuals as well as their rating adjustments
and generates new scores based on the methods of Neumann et al 2011
}
\references{
Neumann et al (2011) Assessing Dominance Hierarchies.
Animal Behaviour
}

