% Generated by roxygen2 (4.0.0): do not edit by hand
\name{subset.interData}
\alias{subset.interData}
\title{Subsets an interData object by players and time}
\usage{
\method{subset}{interData}(intData, players = intData$players,
  time.range = intData$datetime)
}
\arguments{
\item{intData}{an oject of class "interData" to subset.}

\item{players}{players which to restrict the interactions to.}

\item{time.range}{a length 2 sorted POSIXct vector with the first element
specifying the earliest time of which to consider iteractions and the last
element specifying the latest time.}
}
\value{
a "subset" interData object.
}
\description{
Takes an "interData" object and subsets it so that it only includes
interactions involving the specifed players within the specified time range.
}

