% Generated by roxygen2 (4.0.0): do not edit by hand
\name{toBayesDF}
\alias{toBayesDF}
\title{Convert interData object to data frame of wins for individuals}
\usage{
toBayesDF(intData)
}
\arguments{
\item{intData}{object of class "interData" to build the matrix}
}
\value{
a data frame with arow for each dyad in the group indicating number
of wins.
}
\description{
Creates a data frame which has a row for each unique pairing of indvividuals
in a group and a column specifying the number of first player wins as well as
a column for the number of second player wins for that dyad.
}
\details{
Due to the nature of the construction of the matrix ties amongst
dyads are ignored.
}
\examples{
# generate generic data
interactions <- data.frame (a = sample (letters [1:10], 100, T),
                            b = sample (letters [1:10], 100, T),
                            o = sample (c(-1,-1,0,1,1), 100, T),
                            d = Sys.time () + runif (100, 40, 160))
# convert to interData object
id1 <- intTableConv (interactions)
# calculate using all players
toBayesDF (id1)
}

