% Generated by roxygen2 (4.0.0): do not edit by hand
\name{toInterMat}
\alias{toInterMat}
\title{Convert interData object to interaction matrix}
\usage{
toInterMat(intData)
}
\arguments{
\item{intData}{object of class "interData" to build the matrix}
}
\value{
a matrix of player wins
}
\description{
Creates a matrix cataloging the wins between players in
an interData object.
}
\details{
toInterMat creates a matrix which uses players from
intData as row and column names. Each cell is given an integer value
indicating the number of times the player indicated in the row won a dyadic
interaction against the player indicted in the column. Ties are ignored for
this calculation.
}
\examples{
# generate generic data
interactions <- data.frame (a = sample (letters [1:10], 100, T),
                            b = sample (letters [1:10], 100, T),
                            o = sample (c(-1,-1,0,1,1), 100, T),
                            d = Sys.time () + runif (100, 40, 160))
# convert to interData object
id1 <- intTableConv (interactions)
# calculate using all players
toInterMat (id1)
}

