% Generated by roxygen2 (4.0.1): do not edit by hand
\name{upsetElo}
\alias{upsetElo}
\title{Get the biggest upsets from an Elo table}
\usage{
upsetElo(eloTab, N = 10, time.range = eloTab$datetime)
}
\arguments{
\item{eloTab}{an object of class "eloTable" from which ratings are pulled.}

\item{N}{the number of ratings to return.}

\item{time.range}{POSIXct of length two which describes the time range to
look for scores in.}
}
\value{
A data frame sorted by greatest upset
}
\description{
The function returns a data frame with the N biggest upsets in an eloTable
object.
}
\details{
Using an eloTable the function will look for the largest upset
between two players as recorded by their difference in initial score. Only
victories where the unlikely player wins will be considered.
}
\examples{
# generate generic data
interactions <- data.frame (a = sample (letters [1:10], 100, T),
                            b = sample (letters [1:10], 100, T),
                            o = sample (c(-1,-1,0,1,1), 100, T),
                            d = Sys.time () + runif (100, 40, 160))
# convert to interData object
id1 <- intTableConv (interactions)
# produce eloTable object
et1 <- eloTable (id1)
# get top 5 highest scores attained in the table
upsetElo (et1, 5)
}

