% Generated by roxygen2 (4.0.0): do not edit by hand
\name{winStreak}
\alias{winStreak}
\title{Find win streak of players in group}
\usage{
winStreak(intData, players = intData$players, sList = FALSE)
}
\arguments{
\item{intData}{an object of class "interData" from which wins are evaluated.}

\item{players}{teh players to get winstreaks of.}

\item{sList}{logical indicating wether a detailed list should be returned
indicating player records and winstreaks. Default is FALSE.}
}
\value{
data frame detailing the winstreaks of the specified players.
}
\description{
Given an interData object returns the longest win streaks for each player.
}
\details{
The function will return a dataframe with specified players from
the interData object along with their longest win streak and the dates of the
first and last victory.
}
\examples{
# generate generic data
interactions <- data.frame (a = sample (letters [1:10], 100, T),
                            b = sample (letters [1:10], 100, T),
                            o = sample (c(-1,-1,0,1,1), 100, T),
                            d = Sys.time () + runif (100, 40, 160))
# convert to interData object
id1 <- intTableConv (interactions)
# get the winstreaks of players
winStreak (id1)
# with details
winStreak (id1, sList = TRUE)
}

