% Generated by roxygen2 (4.0.0): do not edit by hand
\name{calcW}
\alias{calcW}
\title{Calculate David's score 'w' paramter}
\usage{
calcW(player, intData)
}
\arguments{
\item{player}{in the "interData" object of which to find the 'w' parameter}

\item{intData}{an object of class "interData" which contains the players
interactions.}
}
\description{
Find the 'w' paramter of the David's score algorithim for the associated
player defined in the player argument
}

