% Generated by roxygen2 (4.0.0): do not edit by hand
\name{davidScore}
\alias{davidScore}
\title{Calculate the David's Score of players}
\usage{
davidScore(intData)
}
\arguments{
\item{intData}{object of class "interData" to calculate scores.}
}
\value{
A 2 column data frame specifying the players used in the algorithm
sorted by their corresponding david's score.
}
\description{
Calculates the David's Score of players in an "interData" object.
}
\details{
Using the methods outlined in Gamel et al. 2003 a David's score is
calculated using interactions from intData. Adjusting the players parameter
will modify the algorithm so that only interactions that involve the players
specified will be used.
}
\examples{
# generate generic data
interactions <- data.frame (a = sample (letters [1:10], 100, T),
                            b = sample (letters [1:10], 100, T),
                            o = sample (c(-1,-1,0,1,1), 100, T),
                            d = Sys.time () + runif (100, 40, 160))
# convert to interData object
id1 <- intTableConv (interactions)
# calculate using all players
davidScore (id1)
}
\references{
Gammel et al. (2003) David's Score. Animal Behaviour
}

