% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Dij}
\alias{Dij}
\title{Evaluates David adjusted the probability of Winning}
\usage{
Dij(pi, pj, intData)
}
\arguments{
\item{pi}{player to find win percentage of}

\item{pj}{opposng player}

\item{intData}{object of class "interData" where pi and pj are players}
}
\value{
Numeric value reflecting the percentage of time pi beat pj where
there was a clear winner (i.e. no tie) adjusted for chance.
}
\description{
Calculates the percentage of games player i (pi) won over
player j (pj) across the interData object adjusted fpr chance. Ties are
disregarded in this calculation as per the methods described in Gammel et al.
2003.
}
\details{
Given two players in interaction data finds the David adjusted percentage for
winning of the first player over of the seconed as per Gammel et al. 2003.
}
\examples{
# generate generic data
interactions <- data.frame (a = sample (letters [1:10], 100, T),
                            b = sample (letters [1:10], 100, T),
                            o = sample (c(-1,-1,0,1,1), 100, T),
                            d = Sys.time () + runif (100, 40, 160))
# convert to interData object
id1 <- intTableConv (interactions)
# find Dij betweeen two players in group
Dij (id1$players [1], id1$players [2], id1)
}
\references{
Gammel et al. (2003) David's Score. Animal Behaviour
}

