% Generated by roxygen2 (4.0.0): do not edit by hand
\name{eloTable}
\alias{eloTable}
\title{Generate a series of elo ratings from interaction data}
\usage{
eloTable(intData, initial = 1000, kScaleBounds = c(-Inf, Inf),
  kFactors = 100, minThresh = 100)
}
\arguments{
\item{intData}{An onject of class "interData".}

\item{initial}{The starting score for all participants in the group.}

\item{kScaleBounds}{The player score bins to associate with K factors. See
?findK for more details}

\item{kFactors}{The possible K factors to be used. See ?findK for more
details.}

\item{minThresh}{The minimum value of any players score. See
?eloSingleOutcome for more details}
}
\value{
An object of the class eloTable which contains a vector of players
in the graoup, the time range of interactions, and a data frame sorted by
time containing any changes in an individuals elo rating.
}
\description{
Given an object of the class "interData" will produce a series of elo
ratings up until the latest interaction in the interData object.
}
\details{
Using the methods described in Neumann et al 2011. a series of Elo
ratings are generated from interaction data. Initial scores may either be a
single numeric value or a vector of integers correspong to the players
element of intData. Further arguments are passed to the findK and
singleOutcome functions and include the paramters; kScaleBounds, kFactors,
and minTresh.
}
\examples{
# generate generic data
interactions <- data.frame (a = sample (letters [1:10], 100, T),
                            b = sample (letters [1:10], 100, T),
                            o = sample (c(-1,-1,0,1,1), 100, T),
                            d = Sys.time () + runif (100, 40, 160))
# convert to interData object
id1 <- intTableConv (interactions)
# produce eloTable object
eloTable (id1)
}
\references{
Neumann et al (2011) Assessing Dominance Hierarchies.
Animal Behaviour
}

