% Generated by roxygen2 (4.0.0): do not edit by hand
\name{eloTable}
\alias{eloTable}
\title{Generate a series of elo ratings from interaction data}
\usage{
eloTable(interData, initial = 1000, ...)
}
\arguments{
\item{interData}{an onject of class "interData".}

\item{initial}{the starting score for all participants in the group.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
An object of the class eloTable which contains a list of players
in the graoup, the time range of interactions, and a data frame sorted by
time containing any changes in an individuals elo rating.
}
\description{
Given an object of the class "interData" will produce a series of elo
ratings up until the latest interaction in the interData object.
}
\details{
Using the methods described in Neumann et al 2011. a series of Elo
ratings are generated from interaction data. Initial scores may either be a
single numeric value or a vector of integers correspong to the players
element of interData. Further arguments are passed to the findK and
singleOutcome functions and include the paramters; kScaleBounds, kFactors,
and minTresh.
}
\examples{
# generate generic data
interactions <- data.frame (p1 = c('i', 'j', 'i'), p2 = c('j', 'h', 'h'),
                            o = 1, d = c ('1/1/89', '1/3/89', '1/2/89'))
# convert to interData object
id1 <- intTableConv (interactions, format = '\%m/\%d/\%y')
# produce eloTable object
eloTable (id1)
}
\references{
Neumann et al (2011) Assessing Dominance Hierarchies.
Animal Behaviour
}

