% Generated by roxygen2 (4.0.0): do not edit by hand
\name{findIfWon}
\alias{findIfWon}
\title{Find if player won in interaction}
\usage{
findIfWon(player, intData, row)
}
\arguments{
\item{player}{a player from the row in the interData object}

\item{intData}{an object of class "interData" from which an interaction is
to be evaluated}

\item{row}{the row of the interData data frame to evaluate}
}
\value{
logical indicating the wether the player won or not
}
\description{
Given a row of an interData object determines if the player in question was
the winner.
}
\details{
The function looks at an individual row of a data frame from an
interData object and returns a logical of wether or not the player was
victorious. Will return NA if player is not in row
}

