% Generated by roxygen2 (4.0.0): do not edit by hand
\name{intTableConv}
\alias{intTableConv}
\title{Convert a standard data frame to an interData object}
\usage{
intTableConv(df, ...)
}
\arguments{
\item{df}{data frame with 4 columns as described in the details.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
an object of class interData
}
\description{
The function takes a dataframe with either 4 or 5 columns and converts it to
an interData object to be used for analysis in functions in this package.
}
\details{
The columns of the data frame being passed to this function should
be as follows:

- column 1: a vector with the 1st participant in a dyadic interaction/game

- column 2: a vector with the 2nd participant in a dyadic interaction/game

- column 3: the outcome of the interaction coded as either

    >  1 if player 1 was the winner

    > -1 if player 2 was the winner

    >  0 if the outcome was a tie

- column 4: a vector to be coerced to the class POSIXct denoting the time of
the interaction/game
}
\examples{
# generate generic data
interactions <- data.frame (p1 = c('i', 'j', 'i'), p2 = c('j', 'h', 'h'),
                            o = 1, d = c ('1/1/89', '1/3/89', '1/2/89'))
# convert to interData object
intTableConv (interactions, format = '\%m/\%d/\%y')
}

