% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.eloTable}
\alias{plot.eloTable}
\title{Plot and eloTable object}
\usage{
\method{plot}{eloTable}(eloTab, start.time = eloTab[[2]][1],
  end.time = eloTab[[2]][2], players = eloTab[[1]],
  col = 1:length(eloTab[[1]]), xlab = "Time", ylab = "Rating",
  main = "Elo Rating Across Time", xlim = range(wEloTab$datetime),
  ylim = range(wEloTab$score), lty = 1, lwd = 1, legend = T,
  legend.pos = "topleft", bty = "o", ...)
}
\arguments{
\item{eloTab}{object of the class "eloTable" to build a plot with.}

\item{start.time}{object of the class "POSIXct" for whcih only scores
occuring after this time are graphed.}

\item{end.time}{object of the class "POSIXct" for whcih only scores
occuring before this time are graphed.}

\item{players}{character indicating the players in eloTab to graph.}

\item{col}{vector of length equal to the players argument that
correspond to the colors to be used in the visualization for each player.}

\item{xlab}{a label for the x axis, defaults to "Time".}

\item{ylab}{a label for the y axis, defaults to "Rating".}

\item{main}{a main title for the plot.}

\item{xlim}{the x limits (x1, x2) of the plot. Note that x1 > x2 is allowed
and leads to a ‘reversed axis’.}

\item{ylim}{the y limits of the plot.}

\item{lty,lwd}{the line types and widths for lines appearing in the legend.}

\item{legend}{logical indicating wether to include legend.}

\item{legend.pos}{the position of the legend on the graph}

\item{bty}{the type of box to be drawn around the legend. The allowed values
are "o" (the default) and "n".}

\item{...}{additional parameters to be passed to the base plot function.}
}
\description{
Using the player elo ratings of an "eloTable" object, plots a time series
graph showing changes in individuals scores. Specified players must have
at least one score in the specified time range.
}
\details{
Plotting an eloTable object builds a time series line graph for
each player across the range of dates in the eloTable. The players and
dates used may be adjusted as well as the colors to be used. Additional
paramters may be used to pass arguments to the plot.xy, lines and
legend functions.
}
\examples{
# generate generic data
interactions <- data.frame (a = sample (letters [1:10], 100, T),
                            b = sample (letters [1:10], 100, T),
                            o = sample (c(-1,-1,0,1,1), 100, T),
                            d = Sys.time () + runif (100, 40, 160))
# convert to interData object
id1 <- intTableConv (interactions)
# produce eloTable object
et1 <- eloTable (id1)
# plot all scores
plot (et1)
# plot only three players
plot (et1, players = et1$players [1:3], lty = 3, lwd = 5)
# plot only scores before median
plot (et1, end.time = median (et1$eloTable$datetime, na.rm = T), lty = 6,
      main = 'Elo Rating up to Median Date', legend.pos = "bottomleft")
}

