% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.eloTable}
\alias{print.eloTable}
\title{Print call for eloTable object}
\usage{
\method{print}{eloTable}(eloTab)
}
\arguments{
\item{eloTab}{an object of the class "eloTable"}
}
\description{
Prints each item in EloTable list and head of the elo update table.
}
\details{
Prints only the head of the eloTable data.frame omitting initial
scores
}
\examples{
# generate generic data
interactions <- data.frame (a = sample (letters [1:10], 100, T),
                            b = sample (letters [1:10], 100, T),
                            o = sample (c(-1,-1,0,1,1), 100, T),
                            d = Sys.time () + runif (100, 40, 160))
# convert to interData object
id1 <- intTableConv (interactions)
# produce eloTable object
et1 <- eloTable (id1)
print (et1)
}

