% Generated by roxygen2 (4.0.0): do not edit by hand
\name{singleOutcome}
\alias{singleOutcome}
\title{Generate New Elo Scores from a Single Dyadic Interaction}
\usage{
singleOutcome(RA, RB, outcome, kA, kB, minThresh = 100, ...)
}
\arguments{
\item{RA}{Score of the higher rated player (player A)}

\item{RB}{score of the lower rated player (player B)}

\item{outcome}{either 1 player A wins, -1 player B wins, or 0 tie}

\item{kA}{the k factor for player A}

\item{kB}{the k factor for player B}

\item{minThresh}{the minimum value of any players score}
}
\value{
a vector with player A's and B's new scores
}
\description{
Takes the elo score of two individuals as well as their rating adjustments
and generates new scores based on the methods of Neumann et al 2011
}
\references{
Neumann et al (2011) Assessing Dominance Hierarchies.
Animal Behaviour
}

