% Generated by roxygen2 (4.0.0): do not edit by hand
\name{subset.interData}
\alias{subset.interData}
\title{Subsets an interData object by players and time}
\usage{
\method{subset}{interData}(intData, players = intData$players,
  time.range = intData$datetime, and = TRUE)
}
\arguments{
\item{intData}{an oject of class "interData" to subset.}

\item{players}{players which to restrict the interactions to.}

\item{time.range}{a length 2 sorted POSIXct vector with the first element
specifying the earliest time of which to consider iteractions and the last
element specifying the latest time.}

\item{and}{logical specifying wether the players subsetted should occur in
the player.1 and player.2 position (and = TRUE) or just in a s ingle position
(and = FALSE).}
}
\value{
a "subset" of the interData object.
}
\description{
Takes an "interData" object and subsets it so that it only includes
interactions involving the specifed players within the specified time range.
}
\examples{
# generate generic data
interactions <- data.frame (a = sample (letters [1:10], 100, T),
                            b = sample (letters [1:10], 100, T),
                            o = sample (c(-1,-1,0,1,1), 100, T),
                            d = Sys.time () + runif (100, 40, 160))
# convert to interData object
id1 <- intTableConv (interactions)
# subset by players
subset (id1, id1$players [-c(1,2)])
# get any interaction with a single player
subset (id1, id1$players [1], and = FALSE)
#subset by time
subset (id1, time.range = c(mean (id1$datetime), id1$datetime ['end']))
}

