% Generated by roxygen2 (4.0.1): do not edit by hand
\name{toBayesDF}
\alias{toBayesDF}
\title{Convert interData object to an expanded data frame}
\usage{
toBayesDF(intData)
}
\arguments{
\item{intData}{object of class "interData" to build the matrix}
}
\value{
a data frame with a row for each observation in the group and column
for each player where values of 1 and -1 indicate player 1 and player 2
respectively. The outcome variable is the lat column of the data frame where
1 and 0 indicate where player 1 or player 2 won respectively.
}
\description{
Creates a data frame which has a column for each player in a group
}
\details{
Due to the nature of construction of the data frame ties are
ignored.
}
\examples{
# generate generic data
interactions <- data.frame (a = sample (letters [1:10], 100, T),
                            b = sample (letters [1:10], 100, T),
                            o = sample (c(-1,-1,0,1,1), 100, T),
                            d = Sys.time () + runif (100, 40, 160))
# convert to interData object
id1 <- intTableConv (interactions)
# calculate using only players who had at least 1 win or loss interaction
toBayesDF (id1)
}

